+++ src/middlewares/index.ts
@@ -1,4 +1,3 @@
 export { applyOpenApiMetadata } from './openapi.metadata';
-export { validateRequest } from './openapi.request.validator';
-export { RequestValidator } from './openapi.request.validator.2';
+export { RequestValidator } from './openapi.request.validator';
 export { multipart } from './openapi.multipart';
